+++ tasks.py
@@ -1,359 +1,304 @@
 """Tasks for use with Invoke."""
 
+from distutils.util import strtobool
+from invoke import Collection, task as invoke_task
 import os
-from invoke import task
 
-PYTHON_VER = os.getenv("PYTHON_VER", "3.7")
-NAUTOBOT_VER = os.getenv("NAUTOBOT_VER", "v1.0.0a1")
-NAUTOBOT_SRC_URL = os.getenv("NAUTOBOT_SRC_URL", f"https://github.com/nautobot/nautobot/archive/{NAUTOBOT_VER}.tar.gz")
 
-# Name of the docker image/container
-NAME = os.getenv("IMAGE_NAME", "nautobot-plugin-nornir")
-PWD = os.getcwd()
+def is_truthy(arg):
+    """Convert "truthy" strings into Booleans.
 
-COMPOSE_FILE = "development/docker-compose.yml"
-COMPOSE_OVERRIDE = "docker-compose.override.yml"
-BUILD_NAME = "nautobot_plugin_nornir"
+    Examples:
+        >>> is_truthy('yes')
+        True
+    Args:
+        arg (str): Truthy string (True values are y, yes, t, true, on and 1; false values are n, no,
+        f, false, off and 0. Raises ValueError if val is anything else.
+    """
+    if isinstance(arg, bool):
+        return arg
+    return bool(strtobool(arg))
 
-DEFAULT_ENV = {
-    "NAUTOBOT_VER": NAUTOBOT_VER,
-    "PYTHON_VER": PYTHON_VER,
-    "NAUTOBOT_SRC_URL": NAUTOBOT_SRC_URL,
-}
 
-COMPOSE_APPEND = ""
-if os.path.isfile(COMPOSE_OVERRIDE):
-    COMPOSE_APPEND = f"-f {COMPOSE_OVERRIDE}"
-COMPOSE_COMMAND = f"docker-compose -f {COMPOSE_FILE} {COMPOSE_APPEND} -p {BUILD_NAME}"
+# Use pyinvoke configuration for default values, see http://docs.pyinvoke.org/en/stable/concepts/configuration.html
+# Variables may be overwritten in invoke.yml or by the environment variables INVOKE_NAUTOBOT_PLUGIN_NORNIR_xxx
+namespace = Collection("nautobot_plugin_nornir")
+namespace.configure(
+    {
+        "nautobot_plugin_nornir": {
+            "nautobot_ver": "v1.0.0a1",
+            "project_name": "nautobot_plugin_nornir",
+            "python_ver": "3.6",
+            "local": False,
+            "compose_dir": os.path.join(os.path.dirname(__file__), "development"),
+            "compose_files": ["docker-compose.yml", "../docker-compose.override.yml"],
+        }
+    }
+)
 
-environment = DEFAULT_ENV
-# ------------------------------------------------------------------------------
-# BUILD
-# ------------------------------------------------------------------------------
-@task
-def build(context, nautobot_ver=NAUTOBOT_VER, python_ver=PYTHON_VER):
-    """Build all docker images.
+
+def task(function=None, *args, **kwargs):
+    """Task decorator to override the default Invoke task decorator and add each task to the invoke namespace."""
+
+    def task_wrapper(function=None):
+        """Wrapper around invoke.task to add the task to the namespace as well."""
+        if args or kwargs:
+            task_func = invoke_task(*args, **kwargs)(function)
+        else:
+            task_func = invoke_task(function)
+        namespace.add_task(task_func)
+        return task_func
+
+    if function:
+        # The decorator was called with no arguments
+        return task_wrapper(function)
+    # The decorator was called with arguments
+    return task_wrapper
+
+
+def docker_compose(context, command, **kwargs):
+    """Helper function for running a specific docker-compose command with all appropriate parameters and environment.
 
     Args:
         context (obj): Used to run specific commands
-        nautobot_ver (str): Nautobot version to use to build the container
-        python_ver (str): Will use the Python version docker image to build from
+        command (str): Command string to append to the "docker-compose ..." command, such as "build", "up", etc.
+        **kwargs: Passed through to the context.run() call.
     """
-    DEFAULT_ENV[NAUTOBOT_VER] = nautobot_ver
-    DEFAULT_ENV[PYTHON_VER] = python_ver
+    build_env = {
+        "NAUTOBOT_VER": context.nautobot_plugin_nornir.nautobot_ver,
+        "PYTHON_VER": context.nautobot_plugin_nornir.python_ver,
+    }
+    compose_command = f'docker-compose --project-name {context.nautobot_plugin_nornir.project_name} --project-directory "{context.nautobot_plugin_nornir.compose_dir}"'
+    for compose_file in context.nautobot_plugin_nornir.compose_files:
+        compose_file_path = os.path.join(context.nautobot_plugin_nornir.compose_dir, compose_file)
+        if os.path.isfile(compose_file_path):
+            compose_command += f' -f "{compose_file_path}"'
+    compose_command += f" {command}"
+    print(f'Running docker-compose command "{command}"')
+    return context.run(compose_command, env=build_env, **kwargs)
 
-    context.run(
-        f"{COMPOSE_COMMAND} build",
-        env=DEFAULT_ENV,
-    )
+
+def run_command(context, command, **kwargs):
+    """Wrapper to run a command locally or inside the nautobot container."""
+    if is_truthy(context.nautobot_plugin_nornir.local):
+        context.run(command, **kwargs)
+    else:
+        # Check if netbox is running, no need to start another netbox container to run a command
+        docker_compose_status = "ps --services --filter status=running"
+        results = docker_compose(context, docker_compose_status, hide="out")
+        if "nautobot" in results.stdout:
+            compose_command = f"exec nautobot {command}"
+        else:
+            compose_command = f"run --entrypoint '{command}' nautobot"
+
+        docker_compose(context, compose_command, pty=True)
+
+
+# ------------------------------------------------------------------------------
+# BUILD
+# ------------------------------------------------------------------------------
+@task(
+    help={
+        "force_rm": "Always remove intermediate containers",
+        "cache": "Whether to use Docker's cache when building the image (defaults to enabled)",
+    }
+)
+def build(context, force_rm=False, cache=True):
+    """Build Nautobot docker image."""
+    command = "build"
+
+    if not cache:
+        command += " --no-cache"
+    if force_rm:
+        command += " --force-rm"
+
+    print(f"Building Nautobot with Python {context.nautobot_plugin_nornir.python_ver}...")
+    docker_compose(context, command)
 
 
 # ------------------------------------------------------------------------------
 # START / STOP / DEBUG
 # ------------------------------------------------------------------------------
 @task
-def debug(context, nautobot_ver=NAUTOBOT_VER, python_ver=PYTHON_VER):
-    """Start Nautobot and its dependencies in debug mode.
-
-    Args:
-        context (obj): Used to run specific commands
-        nautobot_ver (str): Nautobot version to use to build the container
-        python_ver (str): Will use the Python version docker image to build from
-    """
-    DEFAULT_ENV[NAUTOBOT_VER] = nautobot_ver
-    DEFAULT_ENV[PYTHON_VER] = python_ver
-
-    print("Starting Nautobot .. ")
-    context.run(
-        f"{COMPOSE_COMMAND} up",
-        env=DEFAULT_ENV,
-    )
+def debug(context):
+    """Start Nautobot and its dependencies in debug mode."""
+    print("Starting Nautobot in debug mode...")
+    docker_compose(context, "up")
 
 
 @task
-def start(context, nautobot_ver=NAUTOBOT_VER, python_ver=PYTHON_VER):
-    """Start Nautobot and its dependencies in detached mode.
-
-    Args:
-        context (obj): Used to run specific commands
-        nautobot_ver (str): Nautobot version to use to build the container
-        python_ver (str): Will use the Python version docker image to build from
-    """
-    DEFAULT_ENV[NAUTOBOT_VER] = nautobot_ver
-    DEFAULT_ENV[PYTHON_VER] = python_ver
-
-    print("Starting Nautobot in detached mode.. ")
-    context.run(
-        f"{COMPOSE_COMMAND} up -d",
-        env=DEFAULT_ENV,
-    )
+def start(context):
+    """Start Nautobot and its dependencies in detached mode."""
+    print("Starting Nautobot in detached mode...")
+    docker_compose(context, "up --detach")
 
 
 @task
-def stop(context, nautobot_ver=NAUTOBOT_VER, python_ver=PYTHON_VER):
-    """Stop Nautobot and its dependencies.
-
-    Args:
-        context (obj): Used to run specific commands
-        nautobot_ver (str): Nautobot version to use to build the container
-        python_ver (str): Will use the Python version docker image to build from
-    """
-    DEFAULT_ENV[NAUTOBOT_VER] = nautobot_ver
-    DEFAULT_ENV[PYTHON_VER] = python_ver
-
-    print("Stopping Nautobot .. ")
-    context.run(
-        f"{COMPOSE_COMMAND} down",
-        env=DEFAULT_ENV,
-    )
+def stop(context):
+    """Stop Nautobot and its dependencies."""
+    print("Stopping Nautobot...")
+    docker_compose(context, "down")
 
 
 @task
-def destroy(context, nautobot_ver=NAUTOBOT_VER, python_ver=PYTHON_VER):
-    """Destroy all containers and volumes.
-
-    Args:
-        context (obj): Used to run specific commands
-        nautobot_ver (str): Nautobot version to use to build the container
-        python_ver (str): Will use the Python version docker image to build from
-    """
-    DEFAULT_ENV[NAUTOBOT_VER] = nautobot_ver
-    DEFAULT_ENV[PYTHON_VER] = python_ver
-
-    context.run(
-        f"{COMPOSE_COMMAND} down",
-        env=DEFAULT_ENV,
-    )
-    context.run(
-        f"docker volume rm -f {BUILD_NAME}_pgdata_nautobot_plugin_nornir",
-        env=DEFAULT_ENV,
-    )
+def destroy(context):
+    """Destroy all containers and volumes."""
+    print("Destroying Nautobot...")
+    docker_compose(context, "down --volumes")
 
 
 # ------------------------------------------------------------------------------
 # ACTIONS
 # ------------------------------------------------------------------------------
 @task
-def nbshell(context, nautobot_ver=NAUTOBOT_VER, python_ver=PYTHON_VER):
-    """Launch a nbshell session.
-
-    Args:
-        context (obj): Used to run specific commands
-        nautobot_ver (str): Nautobot version to use to build the container
-        python_ver (str): Will use the Python version docker image to build from
-    """
-    DEFAULT_ENV[NAUTOBOT_VER] = nautobot_ver
-    DEFAULT_ENV[PYTHON_VER] = python_ver
-
-    context.run(
-        f"{COMPOSE_COMMAND} run nautobot nautobot-server nbshell",
-        env=DEFAULT_ENV,
-        pty=True,
-    )
+def nbshell(context):
+    """Launch an interactive nbshell session."""
+    command = "nautobot-server nbshell"
+    run_command(context, command)
 
 
 @task
-def cli(context, nautobot_ver=NAUTOBOT_VER, python_ver=PYTHON_VER):
-    """Launch a bash shell inside the running Nautobot container.
-
-    Args:
-        context (obj): Used to run specific commands
-        nautobot_ver (str): Nautobot version to use to build the container
-        python_ver (str): Will use the Python version docker image to build from
-    """
-    DEFAULT_ENV[NAUTOBOT_VER] = nautobot_ver
-    DEFAULT_ENV[PYTHON_VER] = python_ver
-
-    context.run(
-        f"{COMPOSE_COMMAND} run nautobot bash",
-        env=DEFAULT_ENV,
-        pty=True,
-    )
+def cli(context):
+    """Launch a bash shell inside the running Nautobot container."""
+    run_command(context, "bash")
 
 
-@task
-def create_user(context, user="admin", nautobot_ver=NAUTOBOT_VER, python_ver=PYTHON_VER):
+@task(
+    help={
+        "user": "name of the superuser to create (default: admin)",
+    }
+)
+def create_user(context, user="admin"):
     """Create a new user in django (default: admin), will prompt for password.
 
     Args:
         context (obj): Used to run specific commands
         user (str): name of the superuser to create
-        nautobot_ver (str): Nautobot version to use to build the container
-        python_ver (str): Will use the Python version docker image to build from
     """
-    DEFAULT_ENV[NAUTOBOT_VER] = nautobot_ver
-    DEFAULT_ENV[PYTHON_VER] = python_ver
-
-    context.run(
-        f"{COMPOSE_COMMAND} run nautobot nautobot-server createsuperuser --username {user}",
-        env=DEFAULT_ENV,
-        pty=True,
-    )
+    command = f"nautobot-server createsuperuser --username {user}"
+    run_command(context, command)
 
 
-@task
-def makemigrations(context, name="", nautobot_ver=NAUTOBOT_VER, python_ver=PYTHON_VER):
-    """Run Make Migration in Django.
-
-    Args:
-        context (obj): Used to run specific commands
-        name (str): Name of the migration to be created
-        nautobot_ver (str): Nautobot version to use to build the container
-        python_ver (str): Will use the Python version docker image to build from
-    """
-    DEFAULT_ENV[NAUTOBOT_VER] = nautobot_ver
-    DEFAULT_ENV[PYTHON_VER] = python_ver
-
-    context.run(
-        f"{COMPOSE_COMMAND} up -d postgres",
-        env=DEFAULT_ENV,
-    )
+@task(
+    help={
+        "name": "name of the migration to be created; if unspecified, will autogenerate a name",
+    }
+)
+def makemigrations(context, name=""):
+    """Perform makemigrations operation in Django."""
+    command = "nautobot-server makemigrations nautobot_plugin_nornir"
 
     if name:
-        context.run(
-            f"{COMPOSE_COMMAND} run nautobot nautobot-server makemigrations nautobot_plugin_nornir --name {name}",
-            env=DEFAULT_ENV,
-        )
-    else:
-        context.run(
-            f"{COMPOSE_COMMAND} run nautobot nautobot-server makemigrations nautobot_plugin_nornir",
-            env=DEFAULT_ENV,
-        )
+        command += f" --name {name}"
 
-    context.run(
-        f"{COMPOSE_COMMAND} down",
-        env=DEFAULT_ENV,
-    )
+    run_command(context, command)
 
 
 # ------------------------------------------------------------------------------
 # TESTS / LINTING
 # ------------------------------------------------------------------------------
+@task(
+    help={
+        "keepdb": "save and re-use test database between test runs for faster re-testing.",
+        "label": "specify a directory or module to test instead of running all Nautobot tests",
+        "failfast": "fail as soon as a single test fails don't run the entire test suite",
+        "buffer": "Discard output from passing tests",
+    }
+)
+def unittest(context, keepdb=False, label="nautobot_plugin_nornir", failfast=False, buffer=True):
+    """Run Nautobot unit tests."""
+    command = f"coverage run --module nautobot.core.cli test {label}"
+
+    if keepdb:
+        command += " --keepdb"
+    if failfast:
+        command += " --failfast"
+    if buffer:
+        command += " --buffer"
+    run_command(context, command)
+
+
 @task
-def unittest(context, nautobot_ver=NAUTOBOT_VER, python_ver=PYTHON_VER):
-    """Run Django unit tests for the plugin.
+def pylint(context):
+    """Run pylint code analysis."""
+    command = 'pylint --init-hook "import nautobot; nautobot.setup()" --rcfile pyproject.toml nautobot_plugin_nornir'
+    run_command(context, command)
 
-    Args:
-        context (obj): Used to run specific commands
-        nautobot_ver (str): Nautobot version to use to build the container
-        python_ver (str): Will use the Python version docker image to build from
-    """
-    DEFAULT_ENV[NAUTOBOT_VER] = nautobot_ver
-    DEFAULT_ENV[PYTHON_VER] = python_ver
 
-    docker = f"{COMPOSE_COMMAND} run nautobot"
-    context.run(
-        f'{docker} sh -c "nautobot-server test nautobot_plugin_nornir"',
-        env=DEFAULT_ENV,
-        pty=True,
-    )
+@task(
+    help={
+        "autoformat": "Apply formatting recommendations automatically, rather than failing if formatting is incorrect.",
+    }
+)
+def black(context, autoformat=False):
+    """Check Python code style with Black."""
+    if autoformat:
+        black_command = "black"
+    else:
+        black_command = "black --check --diff"
+
+    command = f"{black_command} ."
+
+    run_command(context, command)
 
 
 @task
-def pylint(context, nautobot_ver=NAUTOBOT_VER, python_ver=PYTHON_VER):
-    """Run pylint code analysis.
-
-    Args:
-        context (obj): Used to run specific commands
-        nautobot_ver (str): Nautobot version to use to build the container
-        python_ver (str): Will use the Python version docker image to build from
-    """
-    DEFAULT_ENV[NAUTOBOT_VER] = nautobot_ver
-    DEFAULT_ENV[PYTHON_VER] = python_ver
-
-    docker = f"{COMPOSE_COMMAND} run nautobot"
+def pydocstyle(context):
+    """Run pydocstyle to validate docstring formatting adheres to NTC defined standards."""
     # We exclude the /migrations/ directory since it is autogenerated code
-    context.run(
-        f"{docker} sh -c \"cd /source && find . -name '*.py' -not -path '*/migrations/*' | "
-        'PYTHONPATH=/source/development DJANGO_SETTINGS_MODULE=nautobot_config xargs pylint"',
-        env=DEFAULT_ENV,
-        pty=True,
-    )
+    command = "pydocstyle --config=.pydocstyle.ini ."
+    run_command(context, command)
 
 
 @task
-def black(context, nautobot_ver=NAUTOBOT_VER, python_ver=PYTHON_VER):
-    """Run black to check that Python files adhere to its style standards.
-
-    Args:
-        context (obj): Used to run specific commands
-        nautobot_ver (str): Nautobot version to use to build the container
-        python_ver (str): Will use the Python version docker image to build from
-    """
-    DEFAULT_ENV[NAUTOBOT_VER] = nautobot_ver
-    DEFAULT_ENV[PYTHON_VER] = python_ver
-
-    docker = f"{COMPOSE_COMMAND} run nautobot"
-    context.run(
-        f'{docker} sh -c "cd /source && black --check --diff ."',
-        env=DEFAULT_ENV,
-        pty=True,
-    )
+def bandit(context):
+    """Run bandit to validate basic static code security analysis."""
+    command = "bandit --recursive . --configfile .bandit.yml"
+    run_command(context, command)
 
 
 @task
-def pydocstyle(context, nautobot_ver=NAUTOBOT_VER, python_ver=PYTHON_VER):
-    """Run pydocstyle to validate docstring formatting adheres to NTC defined standards.
-
-    Args:
-        context (obj): Used to run specific commands
-        nautobot_ver (str): Nautobot version to use to build the container
-        python_ver (str): Will use the Python version docker image to build from
-    """
-    DEFAULT_ENV[NAUTOBOT_VER] = nautobot_ver
-    DEFAULT_ENV[PYTHON_VER] = python_ver
-
-    docker = f"{COMPOSE_COMMAND} run nautobot"
-    # We exclude the /migrations/ directory since it is autogenerated code
-    context.run(
-        f"{docker} sh -c \"cd /source && find . -name '*.py' -not -path '*/migrations/*' | xargs pydocstyle\"",
-        env=DEFAULT_ENV,
-        pty=True,
-    )
+def flake8(context):
+    """Check for PEP8 compliance and other style issues."""
+    command = "flake8 ."
+    run_command(context, command)
 
 
 @task
-def bandit(context, nautobot_ver=NAUTOBOT_VER, python_ver=PYTHON_VER):
-    """Run bandit to validate basic static code security analysis.
+def yamllint(context):
+    """Run yamllint to validate formating adheres to NTC defined YAML standards.
 
     Args:
         context (obj): Used to run specific commands
-        nautobot_ver (str): Nautobot version to use to build the container
-        python_ver (str): Will use the Python version docker image to build from
     """
-    DEFAULT_ENV[NAUTOBOT_VER] = nautobot_ver
-    DEFAULT_ENV[PYTHON_VER] = python_ver
-
-    docker = f"{COMPOSE_COMMAND} run nautobot"
-    context.run(
-        f'{docker} sh -c "cd /source && bandit --recursive ./ --configfile .bandit.yml"',
-        env=DEFAULT_ENV,
-        pty=True,
-    )
+    command = "yamllint . --format standard"
+    run_command(context, command)
 
 
 @task
-def tests(context, nautobot_ver=NAUTOBOT_VER, python_ver=PYTHON_VER):
-    """Run all tests for this plugin.
+def flake8(context):
+    """Check for PEP8 compliance and other style issues."""
+    command = "flake8 ."
+    run_command(context, command)
 
-    Args:
-        context (obj): Used to run specific commands
-        nautobot_ver (str): Nautobot version to use to build the container
-        python_ver (str): Will use the Python version docker image to build from
-    """
-    DEFAULT_ENV[NAUTOBOT_VER] = nautobot_ver
-    DEFAULT_ENV[PYTHON_VER] = python_ver
 
+@task
+def tests(context):
+    """Run all tests for this plugin."""
     # Sorted loosely from fastest to slowest
     print("Running black...")
-    black(context, nautobot_ver=nautobot_ver, python_ver=python_ver)
+    black(context)
     print("Running bandit...")
-    bandit(context, nautobot_ver=nautobot_ver, python_ver=python_ver)
+    bandit(context)
     print("Running pydocstyle...")
-    pydocstyle(context, nautobot_ver=nautobot_ver, python_ver=python_ver)
+    pydocstyle(context)
+    print("Running yamllint...")
+    yamllint(context)
+    print("Running flake8...")
+    flake8(context)
     print("Running pylint...")
-    pylint(context, nautobot_ver=nautobot_ver, python_ver=python_ver)
+    pylint(context)
     print("Running unit tests...")
-    unittest(context, nautobot_ver=nautobot_ver, python_ver=python_ver)
+    unittest(context)
 
     print("All tests have passed!")