+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## v0.9.4 - 2021-09
+
+### Added
+
+- #25 Switch to GH Actions
+- #26 Fail gracefully from empty queryset and missing platform, update dev env to celery
+
+## v0.9.3 - 2021-07
+
+### Fixed
+
+- #22 Loosen nautobot dep
+
 ## v0.9.2 - 2021-06
 
 ### Added